requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.10.6
